Added initial solution implementation
Modified play() and set_volume_db() functions to add checks for user setting and playing an audio with an undefined (NaN) value. 

If a script provides an undefined value to the set_volume_db() function, the set is not performed, and an error is triggered.

If a script plays a source with the play() function while the player has an undefined volume value, an error is triggered.